Adding tree building capabilities
Addding tree building capabilities, but odesn't keep track of which player asked. So, if 'we' asked, we would know for certain what card was shown, and if not, we wouldn't know for sure. This could mean that the information collected could be stored in the player object, so that we don't accidentally use information from the game we are not allowed. Also, this makes sense because when trying to cheat, we are only ever collecting information (since if a game was played on the comptuer, the random cards dealt could be different than in real life). For that reason, the ClueInfo will be moved to the CluePlayer (or an extending class) for the purposes of 'cheating' at an in person game of clue. So that's what needs to be done next