getAnswer working
Got getAnswer working to solve a ClueSolver tree that is pruned. It does not consider information initally known by the player (cards this player has and doesn't have, as well as center cards). I began to implement this (still doesn't work), but it may be better to split the CLueInfo class into the types of info that can appear for the tree in a refactor.