Added final guess failure deduction
In the rare case that a different player failing a final guess can be used to deduce something, this has been added, making the solver as far as I can tell, perfect. This process is very inefficient for large numbers of possibilities but is very quick since there are never very many final guesses and short circuiting is utilized. The solver in the terminal SolverTreeTest.java now also supports final guesses to be used in deductions.